+++ driverapi/driverapi_test.go
@@ -15,8 +15,8 @@ func TestIPDataMarshalling(t *testing.T) {
 		Pool:         &net.IPNet{IP: net.IP{10, 10, 10, 8}, Mask: net.IPMask{255, 255, 255, 0}},
 		Gateway:      &net.IPNet{IP: net.IP{10, 10, 10, 254}, Mask: net.IPMask{255, 255, 255, 0}},
 		AuxAddresses: map[string]*net.IPNet{
-			"ip1": &net.IPNet{IP: net.IP{10, 10, 10, 1}, Mask: net.IPMask{255, 255, 255, 0}},
-			"ip2": &net.IPNet{IP: net.IP{10, 10, 10, 2}, Mask: net.IPMask{255, 255, 255, 0}},
+			"ip1": {IP: net.IP{10, 10, 10, 1}, Mask: net.IPMask{255, 255, 255, 0}},
+			"ip2": {IP: net.IP{10, 10, 10, 2}, Mask: net.IPMask{255, 255, 255, 0}},
 		},
 	}
 
@@ -59,8 +59,8 @@ func TestValidateAndIsV6(t *testing.T) {
 		Pool:    &net.IPNet{IP: net.IP{10, 10, 10, 8}, Mask: net.IPMask{255, 255, 255, 0}},
 		Gateway: &net.IPNet{IP: net.IP{10, 10, 10, 254}, Mask: net.IPMask{255, 255, 255, 0}},
 		AuxAddresses: map[string]*net.IPNet{
-			"ip1": &net.IPNet{IP: net.IP{10, 10, 10, 1}, Mask: net.IPMask{255, 255, 255, 0}},
-			"ip2": &net.IPNet{IP: net.IP{10, 10, 10, 2}, Mask: net.IPMask{255, 255, 255, 0}},
+			"ip1": {IP: net.IP{10, 10, 10, 1}, Mask: net.IPMask{255, 255, 255, 0}},
+			"ip2": {IP: net.IP{10, 10, 10, 2}, Mask: net.IPMask{255, 255, 255, 0}},
 		},
 	}
 
@@ -87,7 +87,7 @@ func TestValidateAndIsV6(t *testing.T) {
 		t.Fatal(err)
 	}
 	if err = i.Validate(); err == nil {
-		t.Fatalf("expected error but succeded")
+		t.Fatalf("expected error but succeeded")
 	}
 	i.Gateway = nil
 
@@ -96,7 +96,7 @@ func TestValidateAndIsV6(t *testing.T) {
 		t.Fatal(err)
 	}
 	if err = i.Validate(); err == nil {
-		t.Fatalf("expected error but succeded")
+		t.Fatalf("expected error but succeeded")
 	}
 	delete(i.AuxAddresses, "ip2")
 
@@ -105,7 +105,7 @@ func TestValidateAndIsV6(t *testing.T) {
 		t.Fatal(err)
 	}
 	if err = i.Validate(); err == nil {
-		t.Fatalf("expected error but succeded")
+		t.Fatalf("expected error but succeeded")
 	}
 	i.Gateway = nil
 
@@ -114,6 +114,6 @@ func TestValidateAndIsV6(t *testing.T) {
 		t.Fatal(err)
 	}
 	if err = i.Validate(); err == nil {
-		t.Fatalf("expected error but succeded")
+		t.Fatalf("expected error but succeeded")
 	}
 }