+++ libnetwork_internal_test.go
@@ -44,7 +44,7 @@ func TestIpamDriverRegistration(t *testing.T) {
 
 	err = c.(*controller).RegisterIpamDriver("", nil)
 	if err == nil {
-		t.Fatalf("Expected failure, but suceeded")
+		t.Fatalf("Expected failure, but succeeded")
 	}
 	if _, ok := err.(types.BadRequestError); !ok {
 		t.Fatalf("Failed for unexpected reason: %v", err)
@@ -52,7 +52,7 @@ func TestIpamDriverRegistration(t *testing.T) {
 
 	err = c.(*controller).RegisterIpamDriver(ipamapi.DefaultIPAM, nil)
 	if err == nil {
-		t.Fatalf("Expected failure, but suceeded")
+		t.Fatalf("Expected failure, but succeeded")
 	}
 	if _, ok := err.(types.ForbiddenError); !ok {
 		t.Fatalf("Failed for unexpected reason: %v", err)
@@ -72,20 +72,20 @@ func TestNetworkMarshalling(t *testing.T) {
 			netlabel.MacAddress: "a:b:c:d:e:f",
 		},
 		ipamV4Config: []*IpamConf{
-			&IpamConf{
+			{
 				PreferredPool: "10.2.0.0/16",
 				SubPool:       "10.2.0.0/24",
 				Gateway:       "",
 				AuxAddresses:  nil,
 			},
-			&IpamConf{
+			{
 				PreferredPool: "10.2.0.0/16",
 				SubPool:       "10.2.1.0/24",
 				Gateway:       "10.2.1.254",
 			},
 		},
 		ipamV6Config: []*IpamConf{
-			&IpamConf{
+			{
 				PreferredPool: "abcd::/64",
 				SubPool:       "abcd:abcd:abcd:abcd:abcd::/80",
 				Gateway:       "abcd::29/64",
@@ -93,7 +93,7 @@ func TestNetworkMarshalling(t *testing.T) {
 			},
 		},
 		ipamV4Info: []*IpamInfo{
-			&IpamInfo{
+			{
 				PoolID: "ipoolverde123",
 				Meta: map[string]string{
 					netlabel.Gateway: "10.2.1.255/16",
@@ -108,7 +108,7 @@ func TestNetworkMarshalling(t *testing.T) {
 					AuxAddresses: nil,
 				},
 			},
-			&IpamInfo{
+			{
 				PoolID: "ipoolblue345",
 				Meta: map[string]string{
 					netlabel.Gateway: "10.2.1.255/16",
@@ -121,12 +121,12 @@ func TestNetworkMarshalling(t *testing.T) {
 					},
 					Gateway: &net.IPNet{IP: net.IP{10, 2, 1, 254}, Mask: net.IPMask{255, 255, 255, 0}},
 					AuxAddresses: map[string]*net.IPNet{
-						"ip3": &net.IPNet{IP: net.IP{10, 2, 1, 3}, Mask: net.IPMask{255, 255, 255, 0}},
-						"ip5": &net.IPNet{IP: net.IP{10, 2, 1, 55}, Mask: net.IPMask{255, 255, 255, 0}},
+						"ip3": {IP: net.IP{10, 2, 1, 3}, Mask: net.IPMask{255, 255, 255, 0}},
+						"ip5": {IP: net.IP{10, 2, 1, 55}, Mask: net.IPMask{255, 255, 255, 0}},
 					},
 				},
 			},
-			&IpamInfo{
+			{
 				PoolID: "weirdinfo",
 				IPAMData: driverapi.IPAMData{
 					Gateway: &net.IPNet{
@@ -137,7 +137,7 @@ func TestNetworkMarshalling(t *testing.T) {
 			},
 		},
 		ipamV6Info: []*IpamInfo{
-			&IpamInfo{
+			{
 				PoolID: "ipoolv6",
 				IPAMData: driverapi.IPAMData{
 					AddressSpace: "viola",
@@ -343,7 +343,7 @@ func TestAuxAddresses(t *testing.T) {
 
 	for _, i := range input {
 
-		n.ipamV4Config = []*IpamConf{&IpamConf{PreferredPool: i.masterPool, SubPool: i.subPool, AuxAddresses: i.auxAddresses}}
+		n.ipamV4Config = []*IpamConf{{PreferredPool: i.masterPool, SubPool: i.subPool, AuxAddresses: i.auxAddresses}}
 
 		err = n.ipamAllocate()
 
@@ -373,7 +373,7 @@ func TestIpamReleaseOnNetDriverFailures(t *testing.T) {
 
 	// Test whether ipam state release is invoked  on network create failure from net driver
 	// by checking whether subsequent network creation requesting same gateway IP succeeds
-	ipamOpt := NetworkOptionIpam(ipamapi.DefaultIPAM, "", []*IpamConf{&IpamConf{PreferredPool: "10.34.0.0/16", Gateway: "10.34.255.254"}}, nil, nil)
+	ipamOpt := NetworkOptionIpam(ipamapi.DefaultIPAM, "", []*IpamConf{{PreferredPool: "10.34.0.0/16", Gateway: "10.34.255.254"}}, nil, nil)
 	if _, err := c.NewNetwork(badDriverName, "badnet1", ipamOpt); err == nil {
 		t.Fatalf("bad network driver should have failed network creation")
 	}
@@ -397,7 +397,7 @@ func TestIpamReleaseOnNetDriverFailures(t *testing.T) {
 	}
 
 	// Now create good bridge network with different gateway
-	ipamOpt2 := NetworkOptionIpam(ipamapi.DefaultIPAM, "", []*IpamConf{&IpamConf{PreferredPool: "10.34.0.0/16", Gateway: "10.34.255.253"}}, nil, nil)
+	ipamOpt2 := NetworkOptionIpam(ipamapi.DefaultIPAM, "", []*IpamConf{{PreferredPool: "10.34.0.0/16", Gateway: "10.34.255.253"}}, nil, nil)
 	gnw, err = c.NewNetwork("bridge", "goodnet2", ipamOpt2)
 	if err != nil {
 		t.Fatal(err)
@@ -454,4 +454,10 @@ func (b *badDriver) DiscoverDelete(dType discoverapi.DiscoveryType, data interfa
 }
 func (b *badDriver) Type() string {
 	return badDriverName
+}
+func (b *badDriver) ProgramExternalConnectivity(nid, eid string, options map[string]interface{}) error {
+	return nil
+}
+func (b *badDriver) RevokeExternalConnectivity(nid, eid string) error {
+	return nil
 }